+++ src/condor_starter.V6.1/NTsenders.h
@@ -40,7 +40,7 @@ extern "C" {
 	int REMOTE_CONDOR_rename( char *path, char *newpath );
 	int REMOTE_CONDOR_read( int fd, void *data, size_t length );
 	int REMOTE_CONDOR_write( int fd, void *data, size_t length );
-	int REMOTE_CONDOR_lseek( int fd, off_t offset, int whence );
+	off_t REMOTE_CONDOR_lseek( int fd, off_t offset, int whence );
 	int REMOTE_CONDOR_mkdir( char *path, int mode );
 	int REMOTE_CONDOR_rmdir( char *path );
 	int REMOTE_CONDOR_fsync( int fd );
